meta-ostro: initialize combined repo and distro layer
The meta-ostro repositories contains additional files for a merged
repository (like the top-level README) and a distro layer with
configuration files.

The layer currently does not have any recipes, therefore the
corresponding changes to BBPATH are commented out to avoid
warnings from bitbake.

The .templateconf and meta-ostro/conf samples are required to make
oe-init-build-env work in the combined repository. These templates
are from OE-core, with the following changes:
- ostro listed as DISTRO (but not active because not available yet)
- rpm as package class instead of ipk
- all current layers in the combined repository active

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>