ostro-image: better comment about Smack dependencies
The previous wording led to some confusion whether Cynara is faulty
because it does not depend on Smack, in particular because
security-manager was said to link against it.

Iggor suggested to replace "security-manager even links" with
"security-manager instead links" and I am adding further changes to
clarify that the main point is about why the tools get added only when
Smack is active.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>