ostro-image: enable IMA
Enabling the layer enables IMA when compiling the kernel, which is
required for the main change, enabling IMA when building ostro-image.

What that means in practice at the moment is that the files get
signed while creating the rootfs and evmctl gets added to the image
to check this.

Because no IMA policy gets loaded, IMA is not really active and thus
booting and modifying files at runtime works as before.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>