ostro-image: Adjust rootfs size
The images created were just a little over the size of the actual
content. This means there's almost no extra space to copy/create new
content.

This isn't an issue with mkefidisk.sh users since the script is
clever enough to utilize the whole SD/USB media size as the rootfs
partition. However, with 'dd' the resulting .hddimg is copied as is.

Set IMAGE_ROOTFS_EXTRA_SPACE to 512Mbytes to ensure there's enough
additional free space always. The "base" rootfs size is derived from
the initial content size by the build system and IMAGE_ROOTFS_EXTRA_SPACE
is added on top of that.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>