ostro-image: enable whole-disk images on intel-corei7-64, intel-core2-32, intel-quark and qemux86*
intel-corei7-64, intel-core2-32 and intel-quark currently boot with
live images (= hddimg). By stopping to use this unsupported feature we
get rid of the unnecessary "install" option and its dependencies
(grub, parted).

The replacement is a whole-disk image (= hdddirect) with multiple
partitions and (because we enable that) an initramfs. The initramfs is
a custom ostro-initramfs, derived from
core-image-minimal-initramfs so that it uses the initramfs-framework.

This addresses the "need image for VirtualBox" part of IOTOS-467 and
even goes beyond it by also preparing images for VMWare and
qemu/kvm. No further tuning for VirtualBox is planned, so IOTOS-467
can be considered resolved with this.

initramfs-framework supports external modules, which will be used to
initialize IMA in a secure way.

Two qemu targets get also modified, to ensure that the extra features
are also used during development. To use the whole-disk images in
qemu, use:
   runqemu ostro-image qemux86 hdddirect

To speed up compilation, override OSTRO_VM_IMAGE_TYPES in local.conf.
However, at least one of the three default types must be set to get an
hdddirect image. With an empty OSTRO_VM_IMAGE_TYPES, the traditional
"runqemu ext4" will still work, but without the extra features in our
initramfs.

It is possible to use the initramfs together with ext4. An additional
advantage is that this allows passing boot parameters via runqemu:
   # Explicitly mention ostro-initramfs if no image
   # type depends on it.
   bitbake ostro-image ostro-initramfs
   runqemu ostro-image qemux86 ext4 \
           'bootparams=debug rootflags=i_version' \
           'qemuparams=-initrd tmp-glibc/deploy/images/qemux86/ostro-initramfs-qemux86.cpio.gz'

Beware to replicate all relevant boot parameters (see 'bitbake -e
ostro-image | grep -e ^APPEND='). The i_version parameter in the
example above is not needed now, but will be once IMA is active.

Fixes: IOTOS-467
Related-to: IOTOS-619

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>