ostro.conf: turn on Cynara checks in D-Bus daemon
This locks down D-Bus system and session busses such that apps do not
get access to D-Bus services unless the service explicitly declares
that it is ready to deal with talking to less privileged apps.

The feature can be disabled separately in a local.conf, with instructions
given in ostro.conf.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>