ostro-image: get rid of kernel warnings about rootfs mounting
When using ext4 and an initramfs which mounts the rootfs directly, the
kernel has to guess the type of the rootfs, which causes it to print
warnings about not being able to use ext2 and ext3. We can get rid of
those by adding a suitable "rootfstype" boot parameter.

This makes sense regardless of the machine and therefore it is set for
all of them. A BSP or a user in local.conf can override the default,
hence the ?= assignment.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>