ostro-image: mount rootfs read-only
It is better to mount the rootfs read-only initially and only make it
read/write after it has passed an fsck. systemd will do that for us
automatically.

The initial attempt with this (6b9ff50c3e8e2) had to be reverted
(7ec1e98dfd19b6713fb) because our patched e2fsprogs fsck caused disk
corruption. Now we are using a pre-release of 1.43 which should not
have that problem anymore, so we can enable the read-only mounting
again.

Related-To: IOTOS-1282

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>