IMA policy: limit IMA to files owned by root
This mitigates potential issues around performance (when users create
large files under their own uid) and reliability (when users use
libraries like sqlite that do not play well with IMA - see
"[Linux-ima-user] IMA hash update").

This creates another loophole for offline attackers (modify a file,
then change its owner to remove the integrity check for it), but that
attack is already possible anyway (downgrade a signed file to hashed
and create a valid IMA hash), so the security of the system is not
impacted by this change.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>