ostro.conf: Added x11 library requirement overrides
X11 libraries are required to build openjdk, but the libraries are not buildable
if X11 is not present in the DISTRO_FEATURES. Adding X11 however may cause other
issues. Thus the distro check is overriden in the required libraries by
introducing an inc file with the overrides and requiring this file in
ostro.conf.

These changes are purely a workaround for meta-java layer integration without
introducing X11 in the image. If X11 is ever added into the image or if
meta-java layer is no longer required, these changes should be undone.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>