ostro-image: cleaner image class inheritance
"inherit core-image" already looks at variables like IMAGE_FSTYPES to
determine which additional classes it needs to import. Setting
IMAGE_FSTYPES later no longer has the desired effect; the conditional
importing of boot-directdisk.bbclass was a workaround for that.

However, bootimg.bbclass was still getting imported and ran some tasks
(do_bootimg) which may have conflicted with boot-directdisk.bbclass
touching some of the same files, in particular boot loader config
files. This probably caused the "empty boot.conf" problem.

The much simpler fix for the problem is to move the "inherit
core-image" such that IMAGE_FSTYPE changes come before it.

Because syslinux.bbclass also no longer gets inherited in all cases
(boot-directdisk.bbclass is more intelligent in this regard than
bootimg.bbclass), we need to ensure that the "root" boot parameter
gets added when it is not inherited. If it is inherited, we must not
add it, because then it would appear twice (once via syslinux.bbclass,
once via us).

Fixes: IOTOS-936

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>