ostro.conf: blacklist partially broken recipes
Several recipes turned out to have accidental dependencies on MACHINE,
which is problematic when building them for Edison and
intel-core2-32. We do not actually build them right now, so we are
still safe, but they show up in a "world" build and thus break a test
for sstate correctness.

We also do not want to build them accidentally (for example, when
adding a developer image), so black list them for now until the
problem is fixed upstream (for example, by "bitbake.conf: Exclude
sstate-outputdirs flag from checksums" and everything covered by
"oeqa/selftest/sstatetests: Add test that MACHINE doesn't change
target sigs")

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>