Revert "ostro-image: mount rootfs read-only"
This reverts commit 6b9ff50c3e8e222b025de61b52ae841de39d4e1a.

Current fsck.ext4 is not reliable (see
https://bugzilla.yoctoproject.org/show_bug.cgi?id=8544): when updating
the time of the superblock, it damages a file system that otherwise
would be okay. In addition, booting with a wrong system time (older
than image) causes unnecessary reboots when fsck.ext4 updates the file
system and returns with an error code, which causes systemd to reboot.

When mounting read/write (the default before the commit above),
fsck.ext4 cannot run and these problems are avoided. .hddimg images
always have behaved like that.

This is a short-term workaround. The proper solution will be more complicated:
- use a working version of fsck.ext4 (current master is fine)
- ignore system time in fsck.ext4 (broken_system_clock in /etc/e2fsck.conf) and/or
- fix system time properly

Related-to: IOTOS-964
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>