Ostro: Image/filesystem creation/boot mechanism
This patch creates images for EFI devices according to the
method designed to support SW update.

As example, it delivers also a secondary EFI partition, which
is currently just a replica of the main one, but will be used
for storing the maintenance/recovery OS.

As part of the restructuring, the kernel and initramfs are turned
into one single EFI-executable binary, which gets loaded by the
BIOS, instead of the traditional bootloader, which remains only
in the form of a stub placed at the head of the efi binary.

The image is not built using the wic tool (as one might expect)
because wic doesn't support the creation of EFI partitions with
arbitrary content, especially when it comes to bootloaders.
Furthermore, this approach would have also required the creation
on-the-fly of a stepping stone kickstart file, customized for the
specific parameters of the image.

Instead, image creation is implemented through the image-iot class
file, which is capable of producing the image and required partition
in one single step

Upstream-Status: no upstream.

Signed-off-by: Igor Stoppa <igor.stoppa@intel.com>