ostro-image: support images with and without IMA
In the M2 release, IMA could only be disabled in the ostro-image if
the corresponding feature was also disabled in the
ostro-initramfs, thus preventing the building of images with
and without IMA in the same build configuration.

The better solution is to use the same ostro-initramfs for
images with and without IMA and disable IMA via the "noima" boot
parameter.

Including the ima-evm-rootfs.bbclass based on IMAGE_FEATURES is also
problematic. By the time that a derived image gets around to
manipulating IMAGE_FEATURES, the "inherit" statement was already
parsed and evaluated, causing the class to be inherited even in images
where signing is meant to be off.

Instead the class must be disabled after including it. Currently this
is done by prepending a "return" statement to the main function. Support
for such a usage directly by the class would be better and could also
cover the dependency on the native tools.

It is still possible to disable IMA entirely with:
   DISTRO_FEATURES_remove = "ima"
This will also disable IMA in the kernel.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>