imagevariant.bbclass: automatically create image variants
Image variants are images under a different recipe and image name
where certain image features are turned on or off. This is a more
flexible replacement for manually creating image recipes that inherit
the base recipe and then try to make similar modifications.

The advantage is that all variations can be defined in one file in a
single line, which is considerably shorter than having to maintain
many different files with explicit IMAGE_FEATURES_append/remove
entries.

Another advantage is that this can be used to allow users to build
unsupported variations by just adding something to their local.conf.
   In image recipe:
   BBCLASSEXTEND = "${EXTRA_VARIANTS}"
   local.conf:
   EXTRA_VARIANTS = "imagevariants:no-debug-tweaks"

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>