ostro-image.bb: only fix SecurityManager DB when enabled
SecurityManager is expected to be installed if and only if application
privileges are enabled in the image. If not installed, then we can
skip the chown command for its DB file.

Alternatively, the command could always be run and check for the file
instead of failing. But that would no longer detect when creating the
file failed, so the current logic without if check is kept.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>