ostro.conf: Set SDK_UPDATE_URL
Installed eSDK can be updated if SDK_UPDATE_URL is set.
Our CI builds and publishes the eSDKs to a server which
we can use to fetch updates from.

Set SDK_UPDATE_URL to point to the latest build output.
Note: this URL must be updated before making a release.

We abuse the *_PACKAGE_FEED_* URL variables here since
URLs they point to the same place where eSDK is copied.

Fixes: IOTOS-1177

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>