local.conf.sample: switch from rpm to ipk as default package format
We have no technical requirement to use rpm. smart+rpm are a bit more
complete when it comes to on-device package handling, but as we do
not support that, we might as well use ipk. It is considered faster
and more efficient (see http://www.yoctoproject.org/docs/latest/mega-manual/mega-manual.html#ref-classes-package).

Also avoids the rpm dependency on beecrypt, which is not supposed
to be used anymore.

Fixes: IOTOS-922

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>