README.rst: clarify combo-layer setup
The text intentionally only talked about "this repository" to avoid
having to update the text when renaming the repositories. However,
this led to confusion because the README.rst appears in both the
ostro and meta-ostro repos.

Describing maintenance of the combined repository with combo-layer
gave the wrong impression that this is how developers should try out
their changes. To avoid this misconception, also describe two other
approaches (forking ostro and using modified components via
bblayers.conf).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>