base OS: prepare for modifying master definitions
The next step will be to modify the passwd/group.master files. This
will also affect fs-perms.txt, so also set up the layer such that
it uses its own copy of that file.

Right now, they are the same as used in OE-core, i.e. the ones from
Debian base-passwd 3.5.2 and OE-core meta/files/fs-perms.txt.

The .bbappend is licensed under MIT, but the base-passwd files
themselves are under GPLv2+ because that is the license under which
they were copied.

Instead changing the BBPATH to override fs-perms.txt one could also
rename the file and set FILESYSTEM_PERMS_TABLES="ostro-fs-perms.txt".
But I found it easier to correlate that modified file with the
original one when using the same name.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>