base OS: simplify default users and groups
This removes all custom users and groups. The rationale is that the OS
configuration should be minimal, because unused components pose a
(unknown) security risk.

Users and groups that are expected by systemd and udev will be created
based on the systemd-sysusers configurations installed by systemd.

The only difference compared to a pure "systemd" OS is that our
"nobody" user belongs to the "nowhere" group, which might be expected
by some OE-core recipes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>