image-iot.bbclass: do not skip IMAGE_PREPROCESS_COMMANDs
Several operations that are relevant for image creation, in particular
our ima_evm_sign_rootfs but also standard operations like prelinking,
now happen in the new do_image task from OE-core.

We must not override that task and instead have to depend on it,
otherwise IMA signing does not happen and booting fails.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>