ostro-image.bb: read-only /etc/passwd + /etc/groups
With systemd-sysusers.bbclass and the hack for the "nobody" group the
systemd-sysusers service no longer makes any changes to the users and
groups of the system.

We could also remove it and the configuration files entirely to save
some space. For now it is left in the image, which also serves as
a safety check that systemd-sysusers.bbclass really worked.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>