base-passwd: add "nobody" group
systemd-sysusers adds the "nobody" group even when the "nobody" user
already exists. This is probably unnecessary (because the existing
"nobody" user belongs to "nowhere" and does not need the "nobody"
group) and might even be considered a bug.

To avoid modifying /etc/group when systemd-sysusers runs, we pre-define
both "nowhere" and "nobody".

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>