ostro-image.bb: introduce "qatests" image feature
The intention is to make building QA tests optional. Right now,
the node runtime always gets build, even when not needed for the
image, because QA tests listed in EXTRA_IMAGEDEPENDS depend on it.

To achieve this, the ostro-image.bbappend in meta-iotqa will
be changed such that it checks the new feature.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>