ostro-image.bb: add ostro-image-minimal
When working on the core OS, quite often it is useful to ignore
components which are not needed by the core OS. The new
ostro-image-minimal variant is configured like that.

To tweak the content of that image variant,
OSTRO_EXTRA_MINIMAL_IMAGE_FEATURES can be set in a local.conf. For
example, to also remove ima use:
  OSTRO_EXTRA_MINIMAL_IMAGE_FEATURES = "no-ima"

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>