ostro-image.bb: add root auto-login image feature
The new image feature is turned off by default. When activated,
the root user will be logged in automatically by getty on serial
ports and virtual console.

This is achieved on a per-image basis by adding the "--auto-login
root" parameter to the getty invocations in a (configurable) set
of systemd unit files. The advantage of this approach over installing
additional .d unit files is that the ExecStart line does not have to
be duplicated. It's also simpler to implement this way.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>