ostro-image.bb: replace default root password with key-based ssh access
Previously, a root password was always set. This is insecure and
should not be supported explicitly, so we can remove the special code
for it.  In addition, the less secure "debug-tweaks" image feature was
basically always on because we had it in our local.conf.sample.

Now debug-tweaks is commented out (but can still be used), which also
has other positive effects for the default configuration:
- ssh does not accept login with blank password, even if set for
  an account
- post-install debug logging is turned off, which we shouldn't need
  (boot-time post-install execution would be incompatible with the
  read-only rootfs)

Instead, setting OSTRO_ROOT_AUTHORIZED_KEYS can be used to
pre-populate ~root/.ssh/authorized_keys with one or more public
keys. This is less risky than a fixed password because the private
key can be kept secret.

This gets documented in the revised local.conf.sample because
local.conf is what most developers will look at for instructions on
what they can change in a local build.

How to deal with changing ssh host keys is also documented there now.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>