DOC setup sphinx infrastructure and website layout
We're using sphinx to generate a stand-alone website so the documentation
can be read locally with a web browser, and with only some theme magic, can
also be uploaded to the ostro project documentation website.

Sphinx-specific stuff is in the sphinx_build folder, except for the
index.rst file that needs to be up one folder at the documentation "root".

Within the sphinx_build folder is a makefile for linux and a make.bat file
for Windows systems:

   make clean      -- cleans out generated content
   make html       -- generates the html website and stores
                      it in the sphinx_build/_output/html folder
                      with index.html as the home page.

* The _build folder is excluded in the .gitignore file.
* The quick_start folder holds work-in-progress quick start docs and the
  support platforms material was moved into this folder
* architecture/architecture.rst index file created (a placeholder for
  the main architecture doc -TBD, and linking to the other
  architecture docs in that folder)
* howtos/booting-and-installation.md converted to .rst and removed and
  augmented including MinnowBoard material
* howtos/building_images.rst augmented with more material
* howtos/doc-guidelines.rst provides some ReST and Sphinx extentions tips
* COPYING license added (Creative Commons Attribution 4.0 International License)

Signed-off-by: David Kinder <david.b.kinder@intel.com>