ostro-image.bb: add "smack" image feature
Disabling Smack during image creation is more useful than
disabling it for the entire distro, because one can build
images with and without Smack without recompiling.

This was not possible in the past because some recipes (in particular
the modified systemd) were not functional when compiled with Smack
support and then booting without Smack. This has been fixed, so
now security=none works.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>