swupd howto: Make extract_rootfs.sh resilient to parallel runs
If more than one instance of extract_rootfs.sh run in the system then
it's not guarantied that the name of a first available loop device
fetched with `losetup -f` is still available by the time the deivce
gets actually allocated.

So, first try to allocate a loop device and only then infer what is
its name.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@intel.com>