DOC howto organization
Adding some organization to the howto/tech notes so they show up nicely in the
Ostro OS website navigation links.  Also removed the automatic inclusion of
new howtos when they appeared in the folder -- now they need to be explicitly
added to the doc tree (giving us some flexibility for adding material before it's
been fully edited), the same as I was doing for the architecture documents.

Signed-off-by: David Kinder <david.b.kinder@intel.com>