security-threat-analysis.rst: address some review feedback
Damaging the brand in general, not necessarily Intel's...

Attack vector on sensors is the device, not the network.
Securing network access to sensors which bypass the device
are out of scope. Found when adding an explanation about that
last column.

Redundant comma removed.

Add "minimal and secure" network protocols to mitigation
strategies for network attacks.

Exclude hardware-specific attacks from current scope.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>