Fix the emptying of /boot
The uefi application is sufficient for booting, it doesn't need
any of the files that are found in the /boot directory of the rootfs,
and they would use up precious space, if left around.

So kill them. This patch makes it happen as post process of the
do_rootfs task.

Signed-off-by: Igor Stoppa <igor.stoppa@intel.com>