swupd howto: Avoid using cp for extracting rootfs.
It may happen that ACLs get corrupted when copying
EFI and EFI/BOOT directories to rootfs. In order to reduce
failure surface use `rsync -aX` instead of `cp -a` to extract files
from rootfs image.

When we are sure that EFI and EFI/BOOT don't have incomplete
attributes this patch can be reverted.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@intel.com>