local.conf: explicit choice between development and production
When introducing "include ostro-os-production.inc", developers were
instructed to make an explicit choice in their local.conf.  However,
it was possible that a developer remains unaware of that and
unintentionally builds a production image. That would have worked when
building without IMA; with IMA, the missing signing key would have
triggered an error.

Now that choice is enforced with a sanity check: a developer
explicitly has to include ostro-os-production.inc or
ostro-os-development.inc. When using ostro-os-production.inc, IMA
still needs to be configured, which is getting checked more explicitly
by the previous commit.

Building without a change to local.conf bails out immediately with:

   ERROR: local.conf must be explicitly edited to select between building
   production and development images. See the comments in local.conf.sample
   and doc/howtos/building-images.rst.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>