system-and-security-architecture.rst: update root access section
The actual implementation turned out to be slightly different than
originally proposed.

Because this is the architecture document, specific "howto" steps
for adding SSH authorized keys are not provided here. Links need
to be added once there is something to be linked to (specific
locations for those instructions unknown at this time).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>