system-and-security-architecture.rst: avoid using TODO keyword
This documentation is getting published on the web. There are concerns
that documenting gaps in such a visible way may leave a bad
impression, therefore visible TODO remarks get removed and future work
is called out as "will be implemented" instead of "not implemented
yet".

It is okay to keep TODO comments in the .rst files.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>