ostro.conf: set linux-yocto preferred version to 4.1
oe-core/meta-intel recently moved to 4.4 on all MACHINEs we are using.

However, some of the Ostro patches that are applied via linux-yocto_%.bbappend
in meta-ostro-bsp are not (yet) compatible with 4.4. The build fails with
MACHINEs we have not properly set PREFERRED_VERSION to 4.1 for linux-yocto.

Make sure all MACHINEs set PREFERRED_VERSION to 4.1 until we are
ready to move to 4.4.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>