swupd: Switch swupd to use bsdtar instead of GNU tar
Currently GNU tar can't be used when IMA is enabled because
upon extracting a file it first creates it with mknodat(),
sets its xattrs and only then tries to open for writing and
fails since IMA prevents it.

This patch adds patches to swupd making it compatible with
bsdtar and switches it to actually use bsdtar.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>