DOC: include firewall configuration howto in web pages
A firewall-configuration.rst and a link to it in the systemd and
security architecture document were added recently, but without adding
that document to the ones that get converted for the public web pages.

We must include the document to avoid the dangling link and, more
importantly, want it to be visible to everyone.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>