DOC: Hardware Compatibility / Sensors edits
Added organization to the new hardware compatibilty documents and
edited the documents submitted for the wireless-networking
area.  The intent is to organize the HW compatibility docs into logical
groups based on their functionality, with separate docs for each of
the peripherals.  These separate docs would explain details about that module
(as was done for the two Zigbee modules).

at86rf222.rst
cc2520.rst
  These two recently submitted docs were moved down a level in the
  hardware folder hierarchy, along with images they reference.

index.rst
  Added the hardware compability section to the overall documentation index page.

[skip ci]

Signed-off-by: David Kinder <david.b.kinder@intel.com>