local.conf.sample: add section to explain OSTRO_VM_IMAGE_TYPES
Add a section that explains how to use the OSTRO_VM_IMAGE_TYPES to
tune the image formats being generated by bitbake. This commit adds
a little explanation but does not change the default behaviour. The
user will need to uncomment the line and set the variable to his
preference.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>