Ostro CI: provide ostro-image via symlink
The automatic CI testing looks for a compressed "ostro-image" under
various names (depending on the platform), both in the download area
and (in the case of Edison) also inside archives.

We no longer build that and instead want to get
"ostro-image-swupd-dev" tested. Until the CI testing becomes
configurable, provide the expected files via a symlink.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>