ostro: clean up EFI combo update integration and swupd check-update
The EFI combo update works only on dsk images. Therefore it should only get
installed and activated on architectures using the dsk image format, and that
choice should be made in a central place like ostro.conf.

For this to work, the .service file which hooks the efi combo updater
into cannot be in the allarch oe-swupd-helpers, because that must have
the same content regardless of the architecture. It seems simpler to
do it all in a single .bbappend anyway, so now
efi-combo-trigger.service is part of the swupd-client.bbappend, just
like the efi-combo-trigger binary itself.

A side effect is that developers can override that choice by setting
the new OSTRO_USE_DSK_IMAGES, which may be useful for debugging.

Upstream swupd provides a "check-update" systemd unit which runs
regularly to check for updates. In Ostro OS, that unit is not
configured and thus gets disabled entirely.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>