image-dsk.bbclass: include microcode in initrd again
OE-core (rev 00c321996c848) and meta-intel (rev 6effd5d241) recently
switched from using INITRD to INITRD_LIVE. We have to do the same,
otherwise our initrd will not contain the microcode update from
meta-intel, which may be relevant on real hardware.

Better solutions, like using a more generic name like INITRD_HW that
is less specific to the current usage in OE-core, were discussed when
this was discovered (see "[meta-intel] meta-intel.inc: Update INTRD to
INITRD_LIVE"), but are not feasible at this time.

image-dsk.py does not use INITRD anymore, so passing that variable to
it can be removed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>