layer.conf: introduce generic META_OSTRO_BASE
The new variable will be used to find Ostro OS files without having to
add new variables each time. The IMAGE_DSK_BASE is kept because the
class might also be used outside of Ostro OS at some point, but now it
is a normal variables with a default value based on META_OSTRO_BASE.

IMAGE_DSK_BASE is not a good name anymore (was used to find multiple
files, now only image-dsk.py), but fixing that would be an unrelated
change.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>