system-and-security-architecture.rst: update based on recent swupd integration
The number of bundles turned out to be less than expected (basically
one per image and not one per image feature).

swupd currently does not verify signatures itself, so the "TBD" line
for that in the production/development mode comparison can be removed
completely (there is no difference).

We will publish swupd update data also for our ostroproject.org
images.

Finally, a minor language clarification: technically it is possible to
set a root password, it is just not supported at the moment.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>