ostro-os-production.inc: disable consoles/getty and mute logging
This commit adds an image feature called 'muted' that can be
used disable all consoles and login services (getty).

In addition, 'muted' modifies systemd journald's log level such that
only emergency logs are logged. This essentially disables all debug
logs.

This leaves kernel printk setting unchanged. That is a BSP setting
Ostro does not need to touch.

The primary use for 'muted' is an a production image. Therefore, it
is added as a default there.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>