ostro.conf: PNBLACKLIST more recipes
As we are now building the whole world for eSDK, we need to review
what are the recipes that are actually needed/useful.

With meta-openembedded layers (and some others too) we use
whitelist.bbclass to select which recipes are needed. However,
with openembedded-core, PNBLACKLIST is suited better.

This commit blacklists some recipes from openembedded-core that are
not applicable for Ostro.

Some further optimization is still needed. This should give a
safe start.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>