ostro.conf: Modify configurations for new java upstream
Whitelist is updated to account for new recipe names. The x11 hack is
removed as it is no longer required. Recipe blacklist is expanded to
include those recipes that are no longer buildable due to the x11 hack
removal, as well as any recipes that have these recipes as direct or
indirect dependency. Bootstrap VM is changed from jam-vm to cacao-vm, as
jam-vm has a X11 dependency.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>