python-pygobject_%.bbappend: move it into meta-ostro layer
The previous commit should have gone to meta-ostro-fixes. In the
meta-ostro repo the file ended up in a place where it is not used.

The long-term goal is to merge the two repos, so instead of moving the
file to meta-ostro-fixes and back again soon, it simply gets moved
into the "meta-ostro" layer and the layer.conf gets updated such that
it is used there.

The path was chosen to be short (fixes/meta instead of
meta-ostro-fixes/meta-fixes). This is tentative and can be discussed
more widely when also moving the content of the meta-ostro-fixes
content.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>