buildhistory-extra.bbclass: fix storing of extra info
do_populate_sysroot[_setscene] may get skipped entirely if nothing
depends on the files being installed in the sysroot. For example,
adwaita-icon-theme is only needed in the images, so only it's
packaging tests run.

We need to hook into anything that produces binaries during a normal
build, to ensure that we don't miss information about those recipes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>