ostro_security_flags.inc: use SECURITY_STRINGFORMAT
OE-core's security_flags.inc got extended to include format
checking. This broke the SECURITY_CFLAGS_remove exceptions because
removing -Wformat -Wformat-security as part of our exception list
still left -Werror=format-security in the CFLAGS, which is not
accepted by gcc.

Instead we need to do it like security_flags.inc does it an unset the
SECURITY_STRINGFORMAT variable per recipe, which is also a bit shorter
and cleaner.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>