supported-recipes.bbclass: alternative to black/white-listing
Choosing recipes via explicit black- and white-listing turned out to
be cumbersome and limited. A distro using it is hard to modify and
use in ways not envisioned by the distro maintainers.

With this new class, all supported recipes are listed explicitly in
one or more configurable files. World builds get automatically
configured to match that list.

In addition, am optional warning before a build lists unsupported
recipes. It includes instructions for dealing with the warning, for
those developers who are unaware of the mechanism. However, most
normal developers will never see the warning because it is disabled by
default. This is more developer-friendly than the dependency errors
one gets when depending on a blacklisted or not-whitelisted recipe.

The intention is that only developers who add content to a distro use
the warning. In a CI system it can be a fatal error, to catch cases
early where a submitter did not enable some required new components
first.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>