supported-recipes.bbclass: limit dependency checking
An unsupported component deep in the stack can lead to a lot of
output, which is both too verbose and expensive to compute. The new
SUPPORTED_RECIPES_CHECK_DEPENDENCY_LINES limit truncates computation
and thus also the output after a certain number of lines (50 by
default).

The expectation is that users will have a scrollback buffer, so that
they should still get to see the actual error which gets printed
before the dependency information.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>