ostro-image-swupd.bb: add ptest packages to -dev bundle and image
Required by QA. It could also be a separate bundle that is not
installed by default, but then running ptests would depend on swupd
being functional.

It makes more sense to not mix ptest testing with swupd testing. For
the latter, QA is going to define their own dedicated QA bundles with
well-known content that they can test for.

Fixes: IOTOS-1515
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>