image-dsk.bbclass: Revert back to old network interface mode
When attempting to launch .ova that is set into bridged mode, it fails
with an error message. While quickly toggling the mode to nat and then
back to bridged fixes the issue, it still looks bad. As a short term
solution the interface is reverted back to NAT mode until I have time to
investigate what the root cause is.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>