supported recipes: identify recipes by name + collection
We have to detect when recipes come from an unexpected layer because
we are combining layers that were not tested in such combinations
(leading to unexpected results) or may have different priorities (we
prefer OE-core over meta-openembedded, which is the opposite of the
priorities set for these two layers).

Internally, recipes are associated with collections, not layers, so we
have to use the collection name as additional qualifier.

This can be confusing, in particular when layers use rather
unintuitive collection names, but at least we are consistent with how
it is done elsewhere. Implementing our own "file to layer" code would
have issues, like not supporting cases where a developer checks out a
layer into a directory that is not named as expected.

The comments an error message address this difference, so it should be
understandable even to developers not familiar with the collection
concept.

The recommended approach for dealing with recipes from unexpected
layers is BBMASK. Manipulating layer priorities would be a lot more
intrusive, and white/black listing is what we want to get rid of.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>